refactor(api): extract Pydantic schemas to dedicated module
